新しい変換ツールサービスインタフェースと抽象クラスを追加し、シンプルな入出力ファイル変換ツールを実装
