+++ src/main/java/kmg/tool/io/application/service/SimpleOne2OneService.java
@@ -1,6 +1,6 @@
-package kmg.tool.application.service.io;
+package kmg.tool.io.application.service;
 
-import kmg.tool.domain.service.io.One2OneService;
+import kmg.tool.io.domain.service.One2OneService;
 
 /**
  * シンプル1入力ファイルから1出力ファイルへの変換ツールサービスインタフェース<br>
