+++ src/main/java/kmg/tool/io/application/service/impl/AccessorCreationServiceImpl.java
@@ -1,4 +1,4 @@
-package kmg.tool.application.service.io.impl;
+package kmg.tool.io.application.service.impl;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
@@ -13,10 +13,10 @@ import org.springframework.stereotype.Service;
 
 import kmg.core.infrastructure.type.KmgString;
 import kmg.core.infrastructure.types.KmgDelimiterTypes;
-import kmg.tool.application.service.io.AccessorCreationService;
-import kmg.tool.domain.service.io.AbstractInputCsvTemplateOutputProcessorService;
 import kmg.tool.domain.types.KmgToolGenMessageTypes;
 import kmg.tool.infrastructure.exception.KmgToolException;
+import kmg.tool.io.application.service.AccessorCreationService;
+import kmg.tool.io.domain.service.AbstractInputCsvTemplateOutputProcessorService;
 
 /**
  * アクセサ作成サービス<br>
