+++ src/main/java/kmg/tool/io/presentation/ui/cli/AbstractIoTool.java
@@ -1,4 +1,4 @@
-package kmg.tool.presentation.ui.cli.io;
+package kmg.tool.io.presentation.ui.cli;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -8,8 +8,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import kmg.core.domain.service.KmgPfaMeasService;
 import kmg.core.domain.service.impl.KmgPfaMeasServiceImpl;
 import kmg.foundation.infrastructure.context.KmgMessageSource;
-import kmg.tool.domain.service.io.IoService;
 import kmg.tool.domain.types.KmgToolGenMessageTypes;
+import kmg.tool.io.domain.service.IoService;
 import kmg.tool.presentation.ui.cli.AbstractTool;
 
 /**
