+++ src/main/java/kmg/tool/io/presentation/ui/cli/SimpleTwo2OneTool.java
@@ -1,11 +1,11 @@
-package kmg.tool.presentation.ui.cli.io;
+package kmg.tool.io.presentation.ui.cli;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.ConfigurableApplicationContext;
 
-import kmg.tool.application.service.io.SimpleTwo2OneService;
+import kmg.tool.io.application.service.SimpleTwo2OneService;
 
 /**
  * シンプル2入力ファイルから1出力ファイルへの変換ツール
